fix(select): refactor component logic into hooks
