+++ packages/select/src/index.js
@@ -1,6 +1,7 @@
-import { SelectProvider } from "./select-provider"
-import SelectInput from "./select-input"
-import SelectList from "./select-list"
-import SelectItem from "./select-item"
-
-export { SelectProvider as Select, SelectInput, SelectList, SelectItem }
+export { Select, SelectInput, SelectList, SelectOption } from "./select"
+export {
+  useSelect,
+  useSelectInput,
+  useSelectList,
+  useSelectOption,
+} from "./use-select"
